feature: added /privacy-policy.pdf direct path & added href on Footer to download pdf
