--- src/pages/PrivacyPolicy.tsx
@@ -1,241 +1,0 @@
-import { ArrowLeft } from 'lucide-react';
-import { useNavigate } from 'react-router-dom';
-import { Button } from '@/components/ui/button';
-
-const PrivacyPolicy = () => {
-  const navigate = useNavigate();
-
-  return (
-    <div className="min-h-screen bg-background">
-      <div className="max-w-4xl mx-auto px-4 py-8">
-        <Button
-          variant="ghost"
-          onClick={() => navigate('/')}
-          className="mb-8 flex items-center gap-2"
-        >
-          <ArrowLeft className="h-4 w-4" />
-          Back to Home
-        </Button>
-
-        <div className="bg-card rounded-lg shadow-lg p-8">
-          <div className="whitespace-pre-wrap text-foreground text-sm leading-relaxed">
-{`SHILOH CONSULTATIONS
-
-Notice of Privacy Practices
-
-This notice describes how protected health information about you may be used and disclosed and how you can get access to this information. Please review it carefully.
-
-If you have any questions about this notice please contact:
-
-Meredith Arthur, DO
-
-4533 Brambleton Ave, suite 11
-
-Roanoke, VA 24018
-
-https://shilohconsultations.org
-
-(540) 404-2682
-
-It is the policy of Shiloh Consultations to provide you with a privacy notice that explains how your healthcare information is being used or disclosed. Shiloh Consultations is required by law to maintain the privacy of your protected health information and provide a notice of its legal duties and privacy practices with respect to protected health information.
-
-This Notice of Privacy Practices describes how Shiloh Consultations may use and disclose your protected health information to carry out treatment, payment or health care operations and for other purposes that are permitted or required by state or federal law. It also describes your rights to access and control your protected health information. "Protected health information" is information related to your past, present or future physical or mental health or condition and related health care services, including demographics that may identify you.
-
-Shiloh Consultations is required to abide by the terms of this Notice of Privacy Practices currently in effect. We may change the terms of our notice, at any time. The new notice will be effective for all protected health information that we maintain at that time and will be posted at the 4533 Brambleton Avenue, suite 11 office. Upon your request, we will provide you with a revised Notice of Privacy Practices. You may request a revised Notice of Privacy Practices by calling Shiloh Consultations at (540) 404-2682 and requesting that a revised copy be sent to you via email or in the mail. We retain prior versions of the Notice of Privacy Practices for six (6) years from the revision date.
-
-1. USES AND DISCLOSURES OF PROTECTED HEALTH INFORMATION
-
-This Notice of Privacy Practices will tell you the ways in which Shiloh Consultations will use and disclose protected health information about you. We will also describe your rights and certain obligations we have regarding the use and disclosure of protected health information.
-
-For Treatment:
-
-We may use protected health information about you to provide you with medical treatment or services. We may disclose protected health information about you to doctors, nurses, technicians, or other healthcare personnel who are involved in your treatment. For example, a clinician treating you for an infection may need to know if you have another condition that could affect your treatment plan and recovery.
-
-Page 1 of 6
-
-
-Consultations
-
-Consultations also may share protected health information about you in order to coordinate the different things you need, such as authorization review. We also may disclose protected health information about you to people outside the office who may be involved with your overall health care.
-
-For Payment:
-
-We may use and disclose protected health information about you so that the treatment and services you receive may be billed to and payment may be collected from you, an insurance company or a third party. For example, we may use your protected health information from a procedure you received at the clinic so that the department can be reimbursed. We may also use your information to obtain prior approval for a treatment you may receive or to determine whether some other third party will cover the treatment.
-
-For Health Care Operations:
-
-We may use and disclose protected health information about you for health care operations. These uses and disclosures are necessary to make sure all patients receive quality care. For example, we may use protected health information to review your treatment and services and to evaluate the performance of the staff caring for you. We may also combine protected health information about many patients to decide what additional services should be covered, what services are not needed, and whether certain new treatments are effective. We may also disclose information to doctors, nurses, technicians, and other department personnel for review and learning purposes.
-
-Other Permitted and Required Uses and Disclosures That May Be Made Without Your Consent, Authorization or Opportunity to Object
-
-We may use or disclose your protected health information in the following situations without your consent or authorization:
-
-Required By Law:
-
-We may use or disclose your protected health information to the extent that the use or disclosure is required by law. The use or disclosure will be made in compliance with the law and will be limited to the relevant requirements of the law. You will be notified, as required by law, of any such uses or disclosures.
-
-Emergencies:
-
-We may use or disclose your protected health information in an emergency case or situation where it is impractical to obtain your written authorization. If this happens your physician shall try to obtain your oral authorization for a health care provider or health plan to discuss your health records with a third party specified by you.
-
-Public Health:
-
-We may disclose your protected health information for public health activities and purposes to a public health authority that is permitted by law to collect or receive the information. The disclosure will be made for the purpose of controlling disease, injury or disability. We may also disclose your protected health information, if directed by the public health authority, to a foreign government agency that is collaborating with the public health authority.
-
-Communicable Diseases:
-
-We may disclose your protected health information, if authorized by law, to a person who may have been exposed to a communicable disease or may otherwise be at risk of contracting or spreading the disease or condition.
-
-
-
-Health Oversight:
-
-We may disclose protected health information to a health oversight agency for activities authorized by law, such as audits, investigations, and inspections. Oversight agencies seeking this information include government agencies that oversee the health care system, government benefit programs, other government regulatory programs and civil rights laws.
-
-Abuse or Neglect:
-
-We may disclose your protected health information to a public health authority that is authorized by law to receive reports of child abuse or neglect. In addition, we may disclose your protected health information if we believe that you have been a victim of abuse, neglect or domestic violence to the governmental entity or agency authorized to receive such information. In this case, the disclosure will be made consistent with the requirements of applicable federal and state laws.
-
-Food and Drug Administration:
-
-We may disclose your protected health information to a person or company required by the Food and Drug Administration to report adverse events, biologic product deviations, product defects or problems; to track products; to enable product recalls; to make repairs or replacements; or to conduct post marketing surveillance, as required by law.
-
-Legal Proceedings:
-
-We may disclose protected health information in the course of any judicial or administrative proceeding in response to an order of a court or administrative tribunal (to the extent such disclosure is expressly authorized), and in certain conditions in response to a subpoena, search warrant, discovery request or other lawful process.
-
-Law Enforcement:
-
-We may also disclose protected health information, so long as applicable federal and state legal requirements are met, for law enforcement purposes. These law enforcement purposes include (1) legal processes and purposes otherwise required by law, (2) limited information requests for identification and location purposes, (3) evidence of a crime committed on our premises, and (4) suspicion that death has occurred as a result of criminal conduct.
-
-Criminal Activity:
-
-Consistent with applicable federal and state laws, we may disclose your protected health information if you have communicated to your provider a specific and immediate threat to cause serious bodily injury or death to an identifiable person or persons, and your provider believes you have the intent and ability to carry out that threat imminently.
-
-Coroners, Funeral Directors, and Organ Donation:
-
-We may disclose protected health information to a coroner or medical examiner for identification purposes, cause of death determinations or for the coroner or medical examiner to perform other duties authorized by law. We may also disclose protected health information to funeral directors, as authorized by law, in order to carry out funeral-related duties. We may disclose such information in reasonable anticipation of death. Protected health information may be used and disclosed for cadaveric organ, eye or tissue donation purposes.
-
-Research:
-
-We may disclose your protected health information to researchers when an institutional review board that has reviewed the research proposal and established protocols to ensure the privacy of your protected health information has approved their research.
-
-
-
-Military Activity and National Security:
-
-We may use or disclose protected health information as required or authorized by law of individuals who are Armed Forces personnel (1) for activities deemed necessary by appropriate military command authorities; (2) for the purpose of a determination by the Department of Veterans Affairs of your eligibility for benefits; or (3) to foreign military authority if you are a member of the foreign military services. We may also disclose your protected health information to authorized federal officials for conducting national security and intelligence activities, including for the provision of protective services to the President or others legally authorized.
-
-Workers' Compensation:
-
-We may disclose your protected health information as authorized to comply with workers' compensation laws and other similar legally established programs that provides benefits for work-related injuries or illnesses.
-
-Inmates:
-
-We may disclose your protected health information to a correctional institution or in other law enforcement custodial situations if it is necessary for your care, or if the disclosure is required by state or federal law.
-
-Immunization Registry:
-
-We may disclose your immunization history with the Virginia Immunization Information System to help prevent you from receiving unnecessary vaccinations. The Virginia Immunization Information System may disclose child immunization proof to schools.
-
-Business Associates:
-
-Some of our services are provided through contracts or agreement with other public and private entities and some of these contracts or agreements requires that health information be disclosed to the contractor. These contractors are known as "business associates." Examples include physician consultants, laboratories, dentists and lawyers from the Office of the Attorney General. We may disclose your health information to these people so they can perform the job we have asked them to do. Whenever an arrangement between our office and a business associate involves the use or disclosure of your protected health information, we will have a written contract that contains terms that will protect the privacy of your protected health information.
-
-Someone Authorized to Make Decisions on Your Behalf:
-
-We may disclose information to those authorized to make decisions on your behalf, such as a power of attorney or a guardian.
-
-Disclosures Related to Care Provided to Minors:
-
-If you are under the age of 18 and not legally emancipated, we are generally required to disclose your PHI to your parent or guardian. However, as required by Virginia Law (Va. Code § 54.1-2969(E)), we will not reveal any information about treatment you receive for venereal disease or other reportable infectious or contagious disease; pregnancy.
-
-2. OTHER USES OF PROTECTED HEALTH INFORMATION
-
-We will not use or disclose your health information for any purpose other than those identified in the previous sections without your specific, written Authorization. Your health information will not be used and disclosed for marketing purposes or sale without your Authorization. If you give us Authorization to use or disclose health information about you, you may revoke that Authorization, in writing, at any time. If you revoke your Authorization, we will no longer use or disclose information about you for the reasons covered by your written.
-
-
-
-Authorization
-
-but we cannot take back any uses or disclosures already made with your permission. Also, we are required to retain our records of the care we provided to you.
-
-Social Media
-
-Shiloh Consultations uses social media platforms to share information with the public. We will not share any medical information about you on our social media sites without your Authorization. However, if you choose to share information about yourself or others on our social media pages and platforms, please understand that the information is public and will not be protected by privacy laws.
-
-3. YOUR RIGHTS
-
-You have the following rights regarding protected health information we maintain about you:
-
-Right to Inspect and Copy:
-
-You have the right to inspect and copy protected health information that may be used to make decisions about your care. Usually, this includes medical and billing records but does not include psychotherapy notes.
-
-To inspect and copy your protected health information, you must submit your request in writing to Shiloh Consultations at the address on the top of this Notice. If you request a copy of information, we may charge a fee for the cost of copying, mailing or other supplies associated with your request.
-
-We may deny your request to inspect and copy in certain very limited circumstances. If you are denied access to protected health information; you may request the denial be reviewed. For more information call (540) 404-2682.
-
-Right to Amend:
-
-If you feel that protected health information about you is incorrect or incomplete, you may ask us to amend the information. You have the right to request an amendment as long as the information is kept by or for Shiloh Consultations. To request an amendment, your request must be made in writing and submitted to Shiloh Consultations, 4533 Brambleton Ave, suite 11, Roanoke, Virginia 24018 (540) 404-2682. You must provide a reason that supports your request. We may deny your request for an amendment if it is not in writing or does not include a reason to support the request. In addition, we may deny your request if you ask us to amend information that:
-
-- Was not created by us, unless the person or entity that created the information is no longer available to make the amendment;
-- Is not part of the protected health information kept by or for Shiloh Consultations;
-- Is not part of the information which you would be permitted to inspect and copy; or
-- Is accurate and complete.
-
-Right to an Accounting of Disclosures:
-
-You have the right to request an "accounting of disclosures." This is a list of the disclosures we made of protected health information about you. The accounting will not include disclosures:
-
-1. for purposes of treatment, payment, or health care operations;
-2. made to you;
-3. made pursuant to your authorization;
-4. made to friends of family in your presence or because of an emergency or disaster;
-5. for national security or intelligence purposes;
-6. to correctional institutions or law enforcement;
-7. as part of a limited data set; or
-8. incident to otherwise permissible disclosures.
-
-Your request must state a time period for the disclosures, which may not be longer than six (6) years and may not include dates before August 26, 2025. Your request should indicate in what form you want the list (for example, on paper, or electronically).
-
-Page 5 of 6
-
-
-
-Right to Request Restrictions:
-
-You have the right to request a restriction or limitation on the protected health information we use or disclose about you for treatment, payment or health care operations. You also have the right to request a limit on the protected health information we disclose about you to someone who is involved in your care or the payment for your care, like a family member or friend. For example, you can ask that we not use or disclose information about a surgery you had. We are not required to agree to your request. If we do agree, we will comply with your request unless the information is needed to provide you emergency treatment. We are required to agree to your request if you pay for treatment, services, supplies and prescriptions "out of pocket" and you request the information not be communicated to your health plan for payment or health care operations purposes. There may be instances where we are required to release this information if required by law.
-
-To request restrictions, you must make your request in writing to Shiloh Consultations, 4533 Brambleton Ave, suite 11, Roanoke, Virginia 24018. In your request you must tell us (1) what information you want to limit; (2) whether you want to limit our use, disclosure or both; and (3) to whom you want the limits to apply, for example, disclosures to your spouse.
-
-Right to Request Confidential Communications:
-
-You have the right to request that we communicate with you about protected health matters in a certain way or at a certain location. Shiloh Consultations reminds you of upcoming appointments and missed appointments. You can ask that we do not contact you, that we send this correspondence to an address other than your home, or you can ask that we only contact you by phone.
-
-To request confidential communications, you must make your request in writing to Shiloh Consultations. We will not ask you the reason for your request. We will accommodate all reasonable requests. Your request must specify how or where you wish to be contacted.
-
-Right to be Notified of a Breach:
-
-You have the right to be notified in the event that we (or a Business Associate of ours) discover a breach of your unsecured protected health information.
-
-Right to a Paper Copy of this Notice:
-
-You have the right to a paper copy of this notice. You may ask us to give you a copy of this notice at any time. To obtain a paper copy of this notice, go to ShilohConsultations.org or call (540) 404-2682 during regular working hours.
-
-4. COMPLAINTS
-
-If you believe your privacy rights have been violated, you may file a complaint with Shiloh Consultations, 4533 Brambleton Ave, suite 11, Roanoke, VA 24018.
-
-To file a complaint with Shiloh Consultations, you may contact our Privacy Officer at (540) 404-2682 Monday through Friday from 9 AM. to 4:00PM, except holidays or by email to DrArthur@ShilohConsultations.org. You will not be penalized for filing a complaint.
-
-This notice was published and becomes effective on August 26, 2025`}
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default PrivacyPolicy;
